requirements.txt
gdown==4.5.1
pymssql==2.3.0
apache-airflow-providers-common-sql==1.3.1
apache-airflow-providers-microsoft-mssql==1.0.0
pandas==1.5.3
pyodbc==4.0.32


